In Firebase it's common to denormalize data such that you keep a list of references in one location:
->  /users/12345/posts/54321/ => true
    /users/12345/posts/23423/ => true

And data in another:

->  /posts/54321 => {...data...}
    /posts/23432 => {...data...}

It's necessary to perform some data acrobatics to subscribe to the "index"
reference location (/users/...) to get relevant IDs, and then
subscribe to each ID, and compile the results into a structure that's appropriate
for common web application needs: an ordered array of raw javascript objects.

I've written firebase-relational-subscription to help. Specify an index reference and a
data reference, and it will create a subscription object that supplies you with
an up-to-date list.